requirements.txt
Flask==1.1.1
slackclient==2.1.0
python-forecastio==1.4.0
zappa==0.48.2